requirements.txt

# Core API/UI
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
streamlit==1.37.1

# RAG / Embeddings
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==3.0.1

# Parsing
python-docx==1.1.2
pypdf==4.3.1
chardet==5.2.0

# Storage / utils
sqlalchemy==2.0.32
sqlite-utils==3.37
python-dotenv==1.0.1
tqdm==4.66.4
requests==2.32.3

networkx==3.3
pyvis==0.3.2
